Added tools.pl and tools.sh to the directory.
Python is a very powerful and beautiful scripting language, which has
native support for Unicode. Thus, it's best suited for writing all our
Unicode shit, and even all our sorting algorithms.

However, Python pales in comparison when compared to Perl on the grounds
of regexes; Perl has regexes intrinsically built into the language.
Thus, all our regex shit is just *screaming* to be written in Perl.

But having our code written in two different languages presents a
problem: How are we going to make these two pieces communicate? This is
where Bash comes in. With the help of heredocs, we're able to call
different parts of our code, whether they're written in Perl or in
Python. Thus, Bash serves as a "glue" language, to put it like that.

In conclusion, the Python-Perl-Bash trio is the perfect combination for
this project. We should write in each language what is best suited for
the language. "Render therefore unto Caesar the things which are
Caesar's; and unto God the things that are God's."

Thank you for taking the time to read the arguments put forth in this
essay :)